requirements.txt
google-generativeai>=0.3.0
arxiv>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tiktoken>=0.5.0 